모든 id 칼럼 데이터타입 long으로 변경, unsigned로 설정했던 칼럼 데이터타입 수정
이제 알았는데 찾아보니까 자바에서는 왜 unsigned가 안되는거임 ㅡㅡ; 뭔가 했네
역시 C#이 좋아! ^^

+++ src/main/java/net/softbell/bsh/controller/view/MemberView.java
@@ -62,7 +62,7 @@ public class MemberView {//extends ControllerFilter {
     @PostMapping("/signup")
     public String execSignup(MemberDTO memberDto) {
     	// Field
-    	int intResult;
+    	long intResult;
     	
     	// Init
     	intResult = memberService.joinUser(memberDto);
