+++ src/main/java/net/softbell/bsh/domain/entity/GroupPermissionPK.java
@@ -28,8 +28,8 @@ public class GroupPermissionPK implements Serializable
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="node_group_id", insertable=false, updatable=false, unique=true, nullable=false)
-	private int nodeGroupId;
+	private long nodeGroupId;
 
 	@Column(name="member_group_id", insertable=false, updatable=false, unique=true, nullable=false)
-	private int memberGroupId;
+	private long memberGroupId;
 }
